move EventsSchema and types into the analytics package
