+++ apps/analytics/app/api/event/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import insertEvent from "~/db/insertEvent";
-import { EventsSchema } from "~/lib/shared_consts";
+import { EventsSchema } from "@codaco/analytics";
 
 const corsHeaders = {
   "Access-Control-Allow-Origin": "*",
@@ -12,7 +12,6 @@ export const runtime = "edge";
 
 export async function POST(request: NextRequest) {
   const event = (await request.json()) as unknown;
-
   const parsedEvent = EventsSchema.safeParse(event);
 
   if (parsedEvent.success === false) {
@@ -26,6 +25,9 @@ export async function POST(request: NextRequest) {
     const result = await insertEvent({
       ...parsedEvent.data,
       timestamp: new Date(parsedEvent.data.timestamp),
+      message: parsedEvent.data.error?.message,
+      name: parsedEvent.data.error?.name,
+      stack: parsedEvent.data.error?.stack,
     });
     if (result.error) throw new Error(result.error);
